+++ src/index.ts
@@ -7,3 +7,4 @@ export * from "./graceful-shutdown";
 export * from "./store";
 export * from "./log";
 export * from "./quit-prompt";
+export * from "./updater";